// src/store/main/system/type.ts
export interface ISystemState {
  usersList: any[]
  usersTotalCount: number
  roleList: any[]
  roleTotalCount: number
  goodsList: any[]
  goodsTotalCount: number
  menuList: any[]
  menuTotalCount: number
  departmentList: any[]
  departmentTotalCount: number
  categoryList: any[]
  categoryTotalCount: number
  storyList: any[]
  storyTotalCount: number
}